Bakery sale spreadsheet
This is a spreadsheet that consists of 4 questions on the question tab asks for the total revenue for all sales, the average quantity sold per day, which item sold the most and which one sold the least. I validated the total revenue by using =SUM(E2:E31) which came back as  $3491. For the average i entered =AVERAGE(D2:D31) and got 30.63 repeating. When it comes to the min and the max i created a seperate product table and hope with the names of the items. I trained the toys sold by using =SUMIF(B:B,"Brownies",D:D) but i told change the “name” depending on the product. After that i used =SUM(H2:H5) to get the total number of all items sold which was 919. For the max i used =INDEX(G2:G5, MATCH(MAX(H2:H5), H2:H5, 0)) & " - " & MAX(H2:H5) to get the name and number a the specified item which was tiramisu -258 and the min being hot chocolate - 203 using =INDEX(G2:G5, MATCH(MIN(H2:H5), H2:H5, 0)) & " - " & MIN(H2:H5).