differences between spreadsheets and databases
This spreadsheet contains the differences between spreadsheets and databases. I was given 6 questions to be answered, how powerful are they relative to one another, how are each of them used to interact with data, how powerful are the relative to one another, pros and cons when sorting and filtering, pros and cons when sorting and filtering in both spreadsheets and databases.